job notifications no longer override created_at
previously, when the job was processed, the created_at would be set in
that task, and then sent to the persist notification tasks to save in
the database. However, this makes debugging quite difficult as the logs
of when the notification is created doesn't line up with the time it's
created at in the database.

So this removes that from the task (it's an optional field now to ensure
backwards compatibility, but we'll remove that too in time)